+++ character/old/character.js
@@ -1,7 +1,6 @@
 const characters = {
 	old_shixie: ["male", "qun", 3, ["biluan", "lixia"]],
 	panfeng: ["male", "qun", 4, ["kuangfu"]],
-	old_shen_zhaoyun: ["male", "shen", 2, ["oldjuejing", "oldlonghun"], ["shu"]],
 	old_guanyinping: ["female", "shu", 3, ["xueji_old", "oldhuxiao", "oldwuji"]],
 	old_caocao: ["male", "shen", 3, ["junkguixin", "feiying"], ["die:caocao"]],
 	old_chendao: ["male", "shu", 4, ["drlt_wanglie"]],
