+++ character/old/translate.js
@@ -131,12 +131,6 @@ const translates = {
 	old_caochun_prefix: "旧",
 	old_guanyinping: "旧关银屏",
 	old_guanyinping_prefix: "旧",
-	old_shen_zhaoyun: "旧神赵云",
-	old_shen_zhaoyun_prefix: "旧神",
-	oldjuejing: "绝境",
-	oldjuejing_info: "锁定技。①摸牌阶段，你令额定摸牌数+X（X为你已损失的体力值）。②你的手牌上限+2。",
-	oldlonghun: "龙魂",
-	oldlonghun_info: "你可以将花色相同的Y张牌按下列规则使用或打出：♥当【桃】，♦当火【杀】，♣当【闪】，♠当普【无懈可击】（Y为你的体力值且至少为1）。",
 	panfeng: "旧潘凤",
 	panfeng_prefix: "旧",
 	old_shixie: "旧士燮",
