+++ character/sixiang/character.js
@@ -1,0 +1,38 @@
+const characters = {
+	std_sunhao: ["male", "wu", 5, ["stdcanshi", "chouhai", "guiming"], ["zhu"]],
+	std_mateng: ["male", "qun", 4, ["stdxiongyi", "mashu", "stdyouji"], ["zhu"]],
+	std_mayunlu: ["female", "shu", 4, ["stdfengpo", "mashu"]],
+	std_jianggan: ["male", "wei", 3, ["stddaoshu", "stddaizui"]],
+	std_zhouchu: ["male", "wu", 4, ["stdxiongxia"]],
+	std_lvlingqi: ["female", "qun", 4, ["stdhuizhan"]],
+	std_dc_yanghu: ["male", "wei", 4, ["stdmingfa"]],
+	std_dc_luotong: ["male", "wu", 3, ["stdjinjian", "stdrenzheng"]],
+	std_lijue: ["male", "qun", 5, ["stdxiongsuan"]],
+	std_chengpu: ["male", "wu", 4, ["stdchunlao"]],
+	std_db_wenyang: ["male", "wei", 4, ["stdquedi"]],
+	std_re_dengzhi: ["male", "shu", 3, ["stdzhiyinmeng", "stdhehe"]],
+	std_zhangyì: ["male", "shu", 4, ["stdzhiyi"]],
+	std_chengyu: ["male", "wei", 3, ["stdshefu", "stdyibing"]],
+	std_fanyufeng: ["female", "qun", 3, ["stdbazhan", "stdzhanying"]],
+	std_feiyi: ["male", "shu", 3, ["stdtiaohe", "stdqiansu"]],
+
+	std_guanxing: ["male", "shu", 4, ["stdwuyou"]],
+	std_fuhuanghou: ["female", "qun", 3, ["stdqiuyuan", "stdzhuikong"]],
+	std_liubiao: ["male", "qun", 3, ["stdzishou", "zongshi", "stdjujin"], ["zhu"]],
+	std_gongsunyuan: ["male", "qun", 4, ["stdhuaiyi", "stdfengbai"], ["name:公孙|渊", "zhu"]],
+	std_cenhun: ["male", "wu", 3, ["stdjishe", "stdwudu"]],
+	std_simashi: ["male", "wei", 4, ["stdjinglve"], ["name:司马|师"]],
+	std_sunshao: ["male", "wu", 3, ["stddingyi", "stdzuici"]],
+	std_jiangwan: ["male", "shu", 3, ["stdruwu", "stdchengshi"]],
+	std_maliang: ["male", "shu", 3, ["stdxiemu", "stdnaman"]],
+	old_shen_zhaoyun: ["male", "shen", 2, ["oldjuejing", "oldlonghun"], ["shu"]],
+	std_pengyang: ["male", "shu", 3, ["stdxiaofan", "stdtuishi"]],
+	std_xushu: ["male", "shu", 3, ["stdwuyan", "stdjujian"], ["border:wei"]],
+	std_xuezong: ["male", "wu", 3, ["stdfunan", "stdxunjie"]],
+	std_liuzhang: ["male", "qun", 3, ["stdyinge", "stdshiren", "stdjuyi"], ["zhu"]],
+	std_wangyuanji: ["female", "wei", 3, ["stdqianchong", "stdshangjian"]],
+	std_wanglang: ["male", "wei", 3, ["stdgushe", "stdjici"]],
+	std_zhonghui: ["male", "wei", 4, ["stdxingfa"], ["clan:颍川钟氏"]],
+};
+
+export default characters;
