+++ character/sixiang/characterFilter.js
@@ -1,0 +1,5 @@
+import { lib, game, ui, get, ai, _status } from "../../noname.js";
+
+const characterFilters = {};
+
+export default characterFilters;
