+++ character/sixiang/characterReplace.js
@@ -1,0 +1,8 @@
+const characterReplaces = {
+	sunhao: ["sunhao", "std_sunhao"],
+	lvlingqi: ["lvlingqi", "std_lvlingqi"],
+	chengyu: ["chengyu", "std_chengyu"],
+	fanyufeng: ["fanyufeng", "std_fanyufeng"],
+};
+
+export default characterReplaces;
