+++ character/sixiang/dynamicTranslate.js
@@ -1,0 +1,4 @@
+import { lib, game, ui, get, ai, _status } from "../../noname.js";
+
+const dynamicTranslates = {};
+export default dynamicTranslates;
