+++ character/sixiang/intro.js
@@ -1,0 +1,5 @@
+const characterIntro = {
+    std_guanxing: "关兴，名将关羽之子，继承了父亲汉寿亭侯的爵位。年少时即受诸葛亮器重，在蜀汉担任侍中、中监军之职，后在夷陵之战中报了杀父之仇。",
+};
+
+export default characterIntro;
