+++ character/sixiang/perfectPairs.js
@@ -1,0 +1,1 @@
+export default {};
