+++ character/sixiang/sort.js
@@ -1,0 +1,13 @@
+const characterSort = {
+	sixiang_shaoyin: ["std_sunhao", "std_mateng", "std_mayunlu", "std_jianggan", "std_zhouchu", "std_lvlingqi", "std_dc_yanghu", "std_dc_luotong", "std_lijue", "std_chengpu", "std_db_wenyang", "std_re_dengzhi", "std_zhangyì", "std_chengyu", "std_fanyufeng", "std_feiyi"],
+	sixiang_taiyin: ["std_guanxing", "std_fuhuanghou", "std_liubiao", "std_gongsunyuan", "std_cenhun", "std_simashi", "std_sunshao", "std_jiangwan", "std_maliang", "std_xushu", "std_xuezong", "std_liuzhang", "std_wangyuanji", "std_wanglang", "std_zhonghui"],
+	sixiang_trashBin: ["old_shen_zhaoyun", "std_pengyang"],
+};
+
+const characterSortTranslate = {
+	sixiang_shaoyin: "四象封印·少阴",
+	sixiang_taiyin: "四象封印·太阴",
+	sixiang_trashBin: "四象封印·垃圾桶",
+};
+
+export { characterSort, characterSortTranslate };
