+++ character/sixiang/voices.js
@@ -1,0 +1,1 @@
+export default {};
