+++ character/sp/characterReplace.js
@@ -5,14 +5,14 @@ const characterReplaces = {
 	caoang: ["caoang", "yj_caoang", "tw_caoang"],
 	caohong: ["caohong", "tw_re_caohong", "tw_caohong", "yj_caohong"],
 	xiahouba: ["xiahouba", "dc_xiahouba", "tw_xiahouba"],
-	maliang: ["maliang", "re_maliang", "tw_maliang", "ol_maliang", "old_maliang"],
+	maliang: ["maliang", "re_maliang", "tw_maliang", "ol_maliang", "old_maliang", "std_maliang"],
 	dingfeng: ["dingfeng", "tw_dingfeng", "old_dingfeng"],
 	zumao: ["zumao", "tw_zumao"],
 	tw_beimihu: ["beimihu", "tw_beimihu"],
 	panfeng: ["panfeng", "re_panfeng", "std_panfeng"],
 	sunluyu: ["sunluyu", "re_sunluyu", "mb_sunluyu"],
 	jin_simazhao: ["jin_simazhao", "simazhao", "sp_simazhao"],
-	jin_wangyuanji: ["jin_wangyuanji", "wangyuanji", "sp_wangyuanji"],
+	jin_wangyuanji: ["jin_wangyuanji", "wangyuanji", "sp_wangyuanji", "std_wangyuanji"],
 	wangyun: ["clan_wangyun", "wangyun", "dc_wangyun", "re_wangyun", "jsrg_wangyun", "old_wangyun", "pe_wangyun"],
 	zhangliang: ["xin_zhangliang", "re_zhangliang", "zhangliang"],
 	lingju: ["lingju", "old_lingju"],
@@ -43,7 +43,7 @@ const characterReplaces = {
 	furong: ["ol_furong", "furong", "tw_furong"],
 	daxiaoqiao: ["daxiaoqiao", "dc_daxiaoqiao", "tw_daxiaoqiao"],
 	zhugeguo: ["zhugeguo", "tw_zhugeguo", "yue_zhugeguo"],
-	wanglang: ["ol_wanglang", "wanglang", "old_wanglang"],
+	wanglang: ["ol_wanglang", "wanglang", "old_wanglang", "std_wanglang"],
 	tengfanglan: ["tengfanglan", "dc_tengfanglan"],
 	zhangyì: ["ol_zhangyì", "zhangyì", "std_zhangyì"],
 	yuantanyuanshang: ["yuantanyuanshang", "yuantanyuanxiyuanshang"],
@@ -52,7 +52,7 @@ const characterReplaces = {
 	ganfuren: ["ganfuren", "dc_ganfuren", "jsp_ganfuren"],
 	zhouqun: ["ol_zhouqun", "zhouqun"],
 	qianzhao: ["ol_qianzhao", "qianzhao"],
-	ol_pengyang: ["ol_pengyang", "sp_pengyang"],
+	ol_pengyang: ["ol_pengyang", "sp_pengyang", "std_pengyang"],
 	ol_luyusheng: ["ol_luyusheng", "luyusheng"],
 	fanjiangzhangda: ["fanjiangzhangda", "jsrg_fanjiangzhangda"],
 	simalang: ["re_simalang", "simalang"],
