+++ character/standard/skill.js
@@ -2,1057 +2,6 @@ import { lib, game, ui, get, ai, _status } from "../../noname.js";
 
 /** @type { importCharacterConfig['skill'] } */
 const skills = {
-	//四象封印·少阴
-	//孙皓
-	stdcanshi: {
-		audio: "canshi",
-		inherit: "canshi",
-		forced: true,
-		async content(event, trigger, player) {
-			trigger.changeToZero();
-			await player.draw(
-				Math.max(
-					1,
-					game.countPlayer(target => {
-						if (player.hasSkill("guiming") && target != player && target.group == "wu") return true;
-						return target.isDamaged();
-					})
-				)
-			);
-			player.addTempSkill("stdcanshi_effect");
-		},
-		subSkill: {
-			effect: {
-				charlotte: true,
-				trigger: { player: "useCardToPlayered" },
-				filter(event, player) {
-					if (event.card.name != "sha" && get.type(event.card) != "trick") return false;
-					return event.target.isDamaged() && player.countCards("he");
-				},
-				forced: true,
-				autodelay: true,
-				content() {
-					player.chooseToDiscard("he", true);
-				},
-			},
-		},
-	},
-	//马腾
-	stdxiongyi: {
-		unique: true,
-		limited: true,
-		audio: "xiongyi",
-		enable: "phaseUse",
-		filterTarget: true,
-		selectTarget: [1, Infinity],
-		multitarget: true,
-		multiline: true,
-		skillAnimation: true,
-		animationColor: "thunder",
-		async content(event, trigger, player) {
-			player.awakenSkill("stdxiongyi");
-			const targets = event.targets.sortBySeat();
-			let keep = true;
-			while (true) {
-				let stop = false;
-				for (const target of targets) {
-					let next = target
-						.chooseToUse(function (card) {
-							const event = get.event();
-							if (!lib.filter.cardEnabled(card, event.player, event)) return false;
-							return get.name(card) == "sha";
-						}, "雄异：是否使用一张不可被响应的【杀】？")
-						.set("oncard", card => {
-							_status.event.directHit.addArray(game.players);
-						});
-					if (!keep) next.set("prompt2", "若你不使用，则结束此流程");
-					const result = await next.forResult();
-					if (!result.bool && !keep) {
-						stop = true;
-						break;
-					}
-				}
-				if (keep) keep = false;
-				if (stop) break;
-			}
-		},
-		ai: {
-			order: 10,
-			result: {
-				target(player, target) {
-					if (player.hasUnknown()) return 0;
-					return target.countCards("hs");
-				},
-			},
-		},
-	},
-	stdyouji: {
-		audio: 2,
-		trigger: { player: "phaseZhunbeiBegin" },
-		filter(event, player) {
-			return player.canMoveCard(
-				null,
-				true,
-				game.filterPlayer(i => {
-					return i.group == "qun";
-				}),
-				card => {
-					return [3, 4, 6].includes(parseInt(get.subtype(card).slice("equip".length)));
-				},
-				"nojudge"
-			);
-		},
-		direct: true,
-		zhuSkill: true,
-		content() {
-			player
-				.moveCard(
-					game.filterPlayer(i => {
-						return i.group == "qun";
-					}),
-					card => {
-						return [3, 4, 6].includes(parseInt(get.subtype(card).slice("equip".length)));
-					}
-				)
-				.set("prompt", get.prompt2("stdyouji"))
-				.set("nojudge", true)
-				.set("logSkill", "stdyouji");
-		},
-	},
-	//马云禄
-	stdfengpo: {
-		audio: "fengpo",
-		trigger: { source: "damageBegin1" },
-		filter(event, player) {
-			return (
-				event.card.name == "sha" &&
-				[player, event.player].some(target => {
-					return target.isIn() && target.countCards("he");
-				})
-			);
-		},
-		async cost(event, trigger, player) {
-			event.result = await player
-				.chooseTarget(get.prompt2("stdfengpo"), (card, player, target) => {
-					const event = get.event().getTrigger();
-					return [player, event.player]
-						.filter(targetx => {
-							return targetx.isIn() && targetx.countCards("he");
-						})
-						.includes(target);
-				})
-				.set("ai", target => {
-					const player = get.event("player"),
-						aim = get.event().getTrigger().player;
-					let eff = get.damageEffect(aim, player, player);
-					if (aim === player && player.getDiscardableCards(player, "he", card => get.suit(card) == "diamond")) eff /= 4;
-					return eff + get.effect(target, { name: "guohe" }, player, player);
-				})
-				.forResult();
-		},
-		async content(event, trigger, player) {
-			const target = event.targets[0];
-			const result = await player
-				.discardPlayerCard(target, "he", true)
-				.set("ai", button => {
-					const suit = get.suit(button.link);
-					return get.event().att * (suit == "diamond" ? 5 : 1) * get.value(button.link, player);
-				})
-				.set("prompt", "凤魄：弃置" + (target != player ? get.translation(target) : "") + "一张牌")
-				.set("prompt2", "若弃置了方片牌，则此伤害+1")
-				.set("att", get.sgnAttitude(player, target))
-				.forResult();
-			if (result.bool) {
-				if (result.cards && result.cards.some(i => get.suit(i, target) == "diamond")) {
-					player.popup("洗具");
-					trigger.increase("num");
-				}
-			}
-		},
-	},
-	//蒋干
-	stddaoshu: {
-		audio: "daoshu",
-		trigger: { global: "phaseZhunbeiBegin" },
-		filter(event, player) {
-			return game.hasPlayer(target => {
-				return target != event.player && target.countCards("h");
-			});
-		},
-		async cost(event, trigger, player) {
-			event.result = await player
-				.chooseTarget(get.prompt2("stddaoshu"), (card, player, target) => {
-					const event = get.event().getTrigger();
-					return target != event.player && target.countCards("h");
-				})
-				.set("ai", target => {
-					const player = get.event("player");
-					return -1 / target.countCards("h");
-				})
-				.forResult();
-		},
-		async content(event, trigger, player) {
-			player.tempBanSkill("stddaoshu", "roundStart", false);
-			const target = event.targets[0];
-			const result = await player.choosePlayerCard(target, "h", true).forResult();
-			if (result.bool) {
-				const cards = result.cards || [];
-				if (cards.length) {
-					await player.showCards(cards, get.translation(player) + "发动了【盗书】");
-					await trigger.player.gain(cards, target, "give");
-					const suits = cards.reduce((list, card) => {
-						return list.add(get.suit(card, target));
-					}, []);
-					if (suits.length) {
-						for (const i of [player, trigger.player]) {
-							i.addTempSkill("stddaoshu_effect");
-							i.markAuto("stddaoshu_effect", suits);
-						}
-					}
-				}
-			}
-		},
-		subSkill: {
-			effect: {
-				charlotte: true,
-				onremove: true,
-				mod: {
-					cardEnabled(card, player) {
-						if (player.getStorage("stddaoshu_effect").includes(get.suit(card))) return false;
-					},
-					cardSavable(card, player) {
-						if (player.getStorage("stddaoshu_effect").includes(get.suit(card))) return false;
-					},
-				},
-				intro: { content: "不能使用$花色的牌" },
-			},
-		},
-	},
-	stddaizui: {
-		audio: "spdaizui",
-		trigger: { player: "damageEnd" },
-		filter(event, player) {
-			return player.isTempBanned("stddaoshu");
-		},
-		forced: true,
-		content() {
-			delete player.storage.temp_ban_stddaoshu;
-			player.popup("盗书");
-			game.log(player, "重置了技能", "#g【盗书】");
-		},
-		ai: {
-			combo: "stddaoshu"
-		},
-	},
-	//周处
-	stdxiongxia: {
-		audio: "xianghai",
-		enable: "chooseToUse",
-		filterCard: true,
-		selectCard: 2,
-		position: "hes",
-		viewAs: { name: "juedou" },
-		selectTarget: 2,
-		viewAsFilter(player) {
-			if (player.countCards("hes") < 2) return false;
-		},
-		check(card) {
-			if (get.name(card) == "sha") return 4 - get.value(card);
-			return 7.5 - get.value(card);
-		},
-		onuse(links, player) {
-			player.addTempSkill("stdxiongxia_effect");
-		},
-		subSkill: {
-			effect: {
-				charlotte: true,
-				trigger: { player: "useCardAfter" },
-				filter(event, player) {
-					return (
-						event.skill == "stdxiongxia" &&
-						(event.targets || []).every(target => {
-							return target.getHistory("damage", evt => {
-								return evt.card && evt.card == event.card;
-							}).length;
-						})
-					);
-				},
-				forced: true,
-				popup: false,
-				content() {
-					player.tempBanSkill("stdxiongxia");
-				},
-			},
-		},
-	},
-	//吕玲绮
-	stdhuizhan: {
-		audio: "guowu",
-		trigger: { player: "useCard2" },
-		filter(event, player) {
-			if (event.card.name != "sha") return false;
-			return game.hasPlayer(target => {
-				return !event.targets.includes(target) && lib.filter.targetEnabled2(event.card, player, target) && lib.filter.targetInRange(event.card, player, target);
-			});
-		},
-		async cost(event, trigger, player) {
-			event.result = await player
-				.chooseTarget(
-					get.prompt2("stdhuizhan"),
-					(card, player, target) => {
-						const event = get.event().getTrigger();
-						return !event.targets.includes(target) && lib.filter.targetEnabled2(event.card, player, target) && lib.filter.targetInRange(event.card, player, target);
-					},
-					[1, 2]
-				)
-				.set("ai", target => {
-					const player = get.event("player"),
-						event = get.event().getTrigger();
-					return get.effect(target, event.card, player);
-				})
-				.forResult();
-		},
-		async content(event, trigger, player) {
-			trigger.targets.addArray(event.targets);
-			player.addTempSkill("stdhuizhan_effect");
-			trigger.card.stdhuizhan = true;
-		},
-		subSkill: {
-			effect: {
-				charlotte: true,
-				trigger: { global: "chooseToUseBegin" },
-				filter(event, player) {
-					if (event._stdhuizhan_effect) return false;
-					const evt = event.getParent(2);
-					return evt.card && evt.card.stdhuizhan;
-				},
-				forced: true,
-				popup: false,
-				forceDie: true,
-				async content(event, trigger, player) {
-					trigger._stdhuizhan_effect = true;
-					const targets = trigger
-						.getParent(2)
-						.targets.filter(i => {
-							return i != trigger.player;
-						})
-						.sortBySeat();
-					if (targets.length) {
-						for (const target of targets) {
-							if (!target.isIn()) continue;
-							const next = target.chooseToUse("挥战：是否替" + get.translation(trigger.player) + "使用一张【闪】？", { name: "shan" });
-							next.set("ai", () => {
-								const event = _status.event;
-								return get.attitude(event.player, event.source) - 2;
-							});
-							next.set("skillwarn", "替" + get.translation(player) + "打出一张闪");
-							next.autochoose = lib.filter.autoRespondShan;
-							next.set("source", player);
-							const result = await next.forResult();
-							if (result.bool) {
-								trigger.result = { bool: true, card: { name: "shan", isCard: true, cards: result.cards.slice() }, cards: result.cards.slice() };
-								trigger.responded = true;
-								trigger.animate = false;
-								break;
-							}
-						}
-					}
-				},
-			},
-		},
-	},
-	//羊祜
-	stdmingfa: {
-		audio: "dcmingfa",
-		enable: "phaseUse",
-		filter(event, player) {
-			if (player.hasSkill("stdmingfa_used")) return false;
-			return game.hasPlayer(target => target.getHp() > 1);
-		},
-		filterTarget(card, player, target) {
-			return target.getHp() > 1;
-		},
-		async content(event, trigger, player) {
-			const target = event.target;
-			await target.damage();
-			if (target.isIn()) {
-				player.addSkill("stdmingfa_used");
-				player.markAuto("stdmingfa_used", [target]);
-			}
-		},
-		subSkill: {
-			used: {
-				charlotte: true,
-				onremove: true,
-				trigger: { global: ["dieAfter", "recoverAfter"] },
-				filter(event, player) {
-					return player.getStorage("stdmingfa_used").includes(event.player);
-				},
-				forced: true,
-				popup: false,
-				content() {
-					player.popup("明伐");
-					game.log(player, "恢复了技能", "#g【明伐】");
-					player.removeSkill("stdmingfa_used");
-				},
-			},
-		},
-		ai: {
-			order: 10,
-			result: {
-				target(player, target) {
-					return get.sgn(get.attitude(player, target)) * get.damageEffect(target, player, player);
-				},
-			},
-		},
-	},
-	//骆统
-	stdrenzheng: {
-		audio: "renzheng",
-		trigger: { global: ["damageCancelled", "damageZero"] },
-		filter(event, player, name) {
-			if (!_status.currentPhase || !_status.currentPhase.isIn()) return false;
-			if (name == "damageCancelled") return true;
-			return event.change_history.some(i => i < 0);
-		},
-		forced: true,
-		logTarget: () => _status.currentPhase,
-		content() {
-			_status.currentPhase.draw();
-		},
-	},
-	stdjinjian: {
-		audio: "jinjian",
-		trigger: { source: "damageBegin1" },
-		logTarget: "player",
-		filter: function (event, player) {
-			return !event.stdjinjian_source2 && !player.hasSkill("stdjinjian_source2");
-		},
-		prompt2: "令即将对其造成的伤害+1",
-		check: function (event, player) {
-			return (
-				get.attitude(player, event.player) < 0 &&
-				!event.player.hasSkillTag("filterDamage", null, {
-					player: player,
-					card: event.card,
-				})
-			);
-		},
-		usable: 1,
-		content: function () {
-			trigger.stdjinjian_source = true;
-			trigger.num++;
-			player.addTempSkill("stdjinjian_source2");
-		},
-		group: "stdjinjian_player",
-		subSkill: {
-			player: {
-				audio: "jinjian",
-				trigger: { player: "damageBegin4" },
-				filter: function (event, player) {
-					return !event.stdjinjian_player2 && !player.hasSkill("stdjinjian_player2");
-				},
-				prompt2: "令即将受到的伤害-1",
-				usable: 1,
-				content: function () {
-					trigger.stdjinjian_player = true;
-					trigger.num--;
-					player.addTempSkill("stdjinjian_player2");
-				},
-			},
-			source2: {
-				trigger: { source: "damageBegin1" },
-				forced: true,
-				charlotte: true,
-				filter: function (event, player) {
-					return !event.stdjinjian_source;
-				},
-				content: function () {
-					trigger.num--;
-					trigger.stdjinjian_source2 = true;
-					player.removeSkill("stdjinjian_source2");
-				},
-				marktext: " -1 ",
-				intro: {
-					content: "下次造成的伤害-1",
-				},
-			},
-			player2: {
-				trigger: { player: "damageBegin3" },
-				forced: true,
-				charlotte: true,
-				filter: function (event, player) {
-					return !event.stdjinjian_player;
-				},
-				content: function () {
-					trigger.num++;
-					trigger.stdjinjian_player2 = true;
-					player.removeSkill("stdjinjian_player2");
-				},
-				marktext: " +1 ",
-				intro: {
-					content: "下次受到的伤害+1",
-				},
-			},
-		},
-		ai: {
-			maixie_defend: true,
-			threaten: 0.9,
-			effect: {
-				target: function (card, player, target) {
-					if (player.hasSkillTag("jueqing", false, target)) return;
-					if (player._stdjinjian_tmp) return;
-					const count = player.storage.counttrigger;
-					if (count && count.stdjinjian_player && count.stdjinjian_player > 0) return;
-					if (_status.event.getParent("useCard", true) || _status.event.getParent("_wuxie", true)) return;
-					if (get.tag(card, "damage")) {
-						if (target.hasSkill("stdjinjian_player2")) {
-							return [1, -2];
-						} else {
-							if (get.attitude(player, target) > 0) {
-								return [0, 0.2];
-							}
-							if (get.attitude(player, target) < 0 && !player.hasSkillTag("damageBonus")) {
-								var sha = player.getCardUsable({ name: "sha" });
-								player._stdjinjian_tmp = true;
-								var num = player.countCards("h", function (card) {
-									if (card.name == "sha") {
-										if (sha == 0) {
-											return false;
-										} else {
-											sha--;
-										}
-									}
-									return get.tag(card, "damage") && player.canUse(card, target) && get.effect(target, card, player, player) > 0;
-								});
-								delete player._stdjinjian_tmp;
-								if (player.hasSkillTag("damage")) {
-									num++;
-								}
-								if (num < 2) {
-									return [0, 0.8];
-								}
-							}
-						}
-					}
-				},
-			},
-		},
-	},
-	//李傕
-	stdxiongsuan: {
-		audio: "xinfu_langxi",
-		trigger: { player: "phaseZhunbeiBegin" },
-		filter(event, player) {
-			return player.isMaxHp();
-		},
-		async cost(event, trigger, player) {
-			event.result = await player
-				.chooseTarget(
-					"请选择【凶算】的目标",
-					lib.translate.stdxiongsuan_info,
-					(card, player, target) => {
-						return target.getHp() == player.getHp();
-					},
-					[1, Infinity],
-					true
-				)
-				.set("ai", target => {
-					const player = get.event("player");
-					return get.damageEffect(target, player, player);
-				})
-				.forResult();
-		},
-		locked: true,
-		async content(event, trigger, player) {
-			for (const i of event.targets) {
-				await i.damage();
-			}
-		},
-		ai: {
-			effect: {
-				target(card, player, target) {
-					if (
-						target.hp <= 1 ||
-						!target.hasFriend() ||
-						!_status.currentPhase ||
-						!get.tag(card, "damage")
-					) return;
-					let hp = target.hp - 1;
-					if (game.hasPlayer(cur => {
-						return cur.hp > hp;
-					})) return;
-					let ori = game.countPlayer(cur => {
-						return cur.hp === hp + 1 && get.attitude(target, cur) <= 0;
-					}), now = game.countPlayer(cur => {
-						return cur.hp === hp && get.attitude(target, cur) <= 0;
-					}), seat = 1, tar = _status.currentPhase.next;
-					while (tar !== target) {
-						if (get.attitude(target, tar) <= 0) seat++;
-						tar = tar.next;
-					}
-					return [1, 2 * (now - ori) / seat];
-				}
-			}
-		}
-	},
-	//程普
-	stdchunlao: {
-		audio: "chunlao",
-		trigger: { player: "phaseDiscardEnd" },
-		filter(event, player) {
-			return (
-				(event.cards || []).length >= 2 &&
-				game.hasPlayer(target => {
-					return target != player && target.countCards("h");
-				})
-			);
-		},
-		async cost(event, trigger, player) {
-			const cards = trigger.cards;
-			event.result = await player
-				.chooseTarget(get.prompt("stdchunlao"), "用" + get.translation(cards) + "交换一名其他角色的手牌", (card, player, target) => {
-					return target != player && target.countCards("h");
-				})
-				.set("ai", target => {
-					return get.event("cards").length - target.countCards("h") - 0.5;
-				})
-				.set("cards", cards)
-				.forResult();
-		},
-		async content(event, trigger, player) {
-			const cards = trigger.cards,
-				target = event.targets[0];
-			await target.loseToDiscardpile(target.getCards("h"));
-			await target.gain(cards, "gain2").set("giver", player);
-			if (player.isDamaged()) {
-				const bool = await target
-					.chooseBool("是否令" + get.translation(player) + "回复1点体力？")
-					.set("choice", get.recoverEffect(player, target, target) > 0)
-					.forResult("bool");
-				if (bool) {
-					target.line(player);
-					await player.recover();
-				}
-			}
-		},
-	},
-	//文鸯
-	stdquedi: {
-		audio: "dbquedi",
-		enable: "chooseToUse",
-		filterCard: { name: "sha" },
-		position: "hes",
-		viewAs: { name: "juedou" },
-		viewAsFilter(player) {
-			if (!player.countCards("hes", { name: "sha" })) return false;
-		},
-		check(card) {
-			return 6 - get.value(card);
-		},
-	},
-	//邓芝
-	//只因盟
-	stdzhiyinmeng: {
-		audio: "weimeng",
-		trigger: { player: "phaseZhunbeiBegin" },
-		filter(event, player) {
-			return player.countCards("he");
-		},
-		async cost(event, trigger, player) {
-			event.result = await player
-				.chooseCardTarget({
-					prompt: get.prompt2("stdzhiyinmeng"),
-					filterTarget: lib.filter.notMe,
-					filterCard: true,
-					position: "he",
-					selectCard: [1, Infinity],
-					complexCard: true,
-					complexTarget: true,
-					complexSelect: true,
-					ai1(card) {
-						if (ui.selected.cards.length && card.name != "du") return 0;
-						if (card.name == "du") return 114514;
-						return 5 - get.value(card);
-					},
-					ai2(target) {
-						if (!ui.selected.cards.length) return 0;
-						const player = get.event("player"),
-							att = get.attitude(player, target);
-						if (ui.selected.cards[0].name == "du") {
-							if (!target.hasSkillTag("nodu")) return -att;
-							return -0.00001 * att;
-						}
-						return att;
-					},
-				})
-				.forResult();
-		},
-		async content(event, trigger, player) {
-			const target = event.targets[0];
-			await player.give(event.cards, target);
-			await target.chooseToGive("he", [1, Infinity], player);
-		},
-	},
-	stdhehe: {
-		audio: "jianliang",
-		trigger: { player: "phaseDrawEnd" },
-		filter(event, player) {
-			return game.hasPlayer(target => {
-				return target != player && target.countCards("h") == player.countCards("h");
-			});
-		},
-		async cost(event, trigger, player) {
-			event.result = await player
-				.chooseTarget(
-					get.prompt2("stdhehe"),
-					(card, player, target) => {
-						return target != player && target.countCards("h") == player.countCards("h");
-					},
-					[1, 2]
-				)
-				.set("ai", target => {
-					const player = get.event("player");
-					return get.effect(target, { name: "draw" }, player, player);
-				})
-				.forResult();
-		},
-		locked: true,
-		async content(event, trigger, player) {
-			await game.asyncDraw(event.targets);
-			await game.delayx();
-		},
-	},
-	//张翼
-	stdzhiyi: {
-		audio: "zhiyi",
-		trigger: { global: "phaseEnd" },
-		filter(event, player) {
-			return player.getHistory("useCard", evt => {
-				return evt.card.name == "sha";
-			}).length;
-		},
-		forced: true,
-		async content(event, trigger, player) {
-			const result = await player.chooseUseTarget("执义：视为使用【杀】，或摸一张牌", { name: "sha" }, false).forResult();
-			if (!result.bool) await player.draw();
-		},
-	},
-	//大魏汉尼拔
-	stdshefu: {
-		audio: "shefu",
-		trigger: { player: "phaseJieshuBegin" },
-		filter(event, player) {
-			return player.countCards("h");
-		},
-		async cost(event, trigger, player) {
-			event.result = await player
-				.chooseCard(get.prompt("stdshefu"), "将一张手牌置于武将牌上", "h")
-				.set("ai", card => {
-					return (
-						(lib.card.list
-							.slice()
-							.map(list => list[2])
-							.filter(name => {
-								return card.name == name;
-							}).length -
-							1) /
-						(get.value(card) || 0.5)
-					);
-				})
-				.forResult();
-		},
-		content() {
-			player.addToExpansion(event.cards, player, "giveAuto").gaintag.add("stdshefu");
-		},
-		marktext: "伏",
-		intro: {
-			markcount: "expansion",
-			mark(dialog, _, player) {
-				const cards = player.getExpansions("stdshefu");
-				if (player.isUnderControl(true) && cards.length) dialog.addAuto(cards);
-				else return "共有" + get.cnNumber(cards.length) + "张“伏兵”";
-			},
-		},
-		onremove(player, skill) {
-			const cards = player.getExpansions(skill);
-			if (cards.length) player.loseToDiscardpile(cards);
-		},
-		group: "stdshefu_effect",
-		subSkill: {
-			effect: {
-				audio: "shefu",
-				trigger: { global: "useCard" },
-				filter(event, player) {
-					return player.getExpansions("stdshefu").some(card => card.name == event.card.name);
-				},
-				async cost(event, trigger, player) {
-					let result = await player
-						.chooseButton(["###" + get.prompt("stdshefu") + "###弃置一张同名牌，令此牌无效", player.getExpansions("stdshefu")])
-						.set("filterButton", button => {
-							return button.link.name == get.event().getTrigger().card.name;
-						})
-						.set("ai", button => {
-							return get.event("goon") ? 1 : 0;
-						})
-						.set("goon", lib.skill.sbkanpo.subSkill.kanpo.check(trigger, player))
-						.forResult();
-					if (result.bool && result.links) {
-						result.cards = result.links.slice();
-						delete result.links;
-					}
-					event.result = result;
-				},
-				async content(event, trigger, player) {
-					await player.loseToDiscardpile(event.cards);
-					trigger.targets.length = 0;
-					trigger.all_excluded = true;
-				},
-			},
-		},
-	},
-	stdyibing: {
-		audio: "benyu",
-		trigger: { global: "dying" },
-		filter(event, player) {
-			return event.player != player && event.player.countCards("h");
-		},
-		direct: true,
-		content() {
-			const target = trigger.player;
-			player.gainPlayerCard(target, "h", true).set("prompt", get.prompt("stdyibing", target)).logSkill = ["stdyibing", target];
-		},
-	},
-	//樊玉凤
-	stdbazhan: {
-		audio: "bazhan",
-		enable: "phaseUse",
-		filter(event, player) {
-			return player.countCards("h") > 0;
-		},
-		filterCard: true,
-		position: "h",
-		filterTarget(card, player, target) {
-			return target != player && target.hasSex("male");
-		},
-		discard: false,
-		lose: false,
-		delay: false,
-		usable: 1,
-		check(card) {
-			if (card.name == "du") return 114514;
-			return 5 - get.value(card);
-		},
-		async content(event, trigger, player) {
-			const target = event.target;
-			await player.give(event.cards, target, "visible");
-			await target
-				.chooseToGive(
-					player,
-					(card, player) => {
-						return get.type2(card) != get.type2(get.event("cards")[0]);
-					},
-					"he"
-				)
-				.set("cards", event.cards);
-		},
-		ai: {
-			order: 10,
-			result: {
-				target(player, target) {
-					if (!ui.selected.cards.length) return 0;
-					const cardxx = ui.selected.cards[0];
-					if (cardxx.name == "du") return -100;
-					if (!player.hasSkill("stdzhanying")) return 1;
-					if (target.countMark("stdzhanying_count") == target.countCards("h") + 1) {
-						const cards = player.getCards("hs", card => {
-							return card != cardxx && get.tag(card, "damage") && player.canUse(card, target) && get.effect(target, card, player, player) > 0;
-						});
-						if (!cards.length) return 1;
-						let cardx = cards.filter(card => get.name(card) == "sha");
-						cardx.sort((a, b) => get.effect(target, b, player, player) - get.effect(target, a, player, player));
-						cardx = cardx.slice(Math.min(cardx.length, player.getCardUsable("sha")), cardx.length);
-						cards.removeArray(cardx);
-						return -cards.reduce((sum, card) => sum + get.effect(target, card, player, player), 0);
-					}
-					return 1;
-				},
-			},
-		},
-	},
-	stdzhanying: {
-		audio: "jiaoying",
-		trigger: { global: "damageBegin2" },
-		filter(event, player) {
-			if (_status.currentPhase !== player) return false;
-			return event.player.countCards("h") > event.player.countMark("stdzhanying_count");
-		},
-		forced: true,
-		logTarget: "player",
-		content() {
-			trigger.increase("num");
-		},
-		global: "stdzhanying_mark",
-		subSkill: {
-			count: {
-				charlotte: true,
-				onremove: true,
-				intro: {
-					markcount: storage => (storage || 0).toString(),
-					content: "本回合开始时手牌数为#张",
-				},
-			},
-			mark: {
-				charlotte: true,
-				trigger: { global: "phaseBegin" },
-				filter(event, player) {
-					return event.player.hasSkill("stdzhanying", null, null, false);
-				},
-				forced: true,
-				popup: false,
-				firstDo: true,
-				content() {
-					player.addTempSkill("stdzhanying_count");
-					player.addMark("stdzhanying_count", player.countCards("h"), false);
-				},
-				mod: {
-					cardEnabled(card, player) {
-						if (!_status.currentPhase || !_status.currentPhase.hasSkill("stdzhanying")) return;
-						if (get.color(card) == "red" && player.countMark("stdzhanying_count") < player.countCards("h")) return false;
-					},
-					cardSavable(card, player) {
-						if (!_status.currentPhase || !_status.currentPhase.hasSkill("stdzhanying")) return;
-						if (get.color(card) == "red" && player.countMark("stdzhanying_count") < player.countCards("h")) return false;
-					},
-				},
-			},
-		},
-	},
-	//F1
-	stdtiaohe: {
-		audio: "fyjianyu",
-		enable: "phaseUse",
-		usable: 1,
-		filter(event, player) {
-			return game.hasPlayer(tar1 => {
-				return tar1.countDiscardableCards(player, "e", i => get.subtype(i) == "equip2") && game.hasPlayer(tar2 => {
-					return tar1 !== tar2 && tar2.countDiscardableCards(player, "e");
-				});
-			});
-			let e = 0, fj = false;
-			game.countPlayer(target => {
-				let es = target.getDiscardableCards(player, "e"), js = target.getDiscardableCards(player, "j", i => get.type(i) == "equip");
-				if (es.length) e++;
-				e += js.length;
-				if (!fj && (es.some(card => get.subtype(card) == "equip2") || js.some(card => get.subtype(card) == "equip2"))) fj = true;
-			});
-			return fj && e >= 2;
-		},
-		filterTarget(card, player, target) {
-			if (!ui.selected.targets.length || ui.selected.targets[0].countDiscardableCards(player, "e", i => get.subtype(i) == "equip2")) {
-				return target.countDiscardableCards(player, "e");
-			}
-			return target.countDiscardableCards(player, "e", i => get.subtype(i) == "equip2");
-			let e = 0;
-			let es = target.getDiscardableCards(player, "e"), js = target.getDiscardableCards(player, "j", i => get.type(i) == "equip");
-			if (es.length) e++;
-			e += js.length;
-			if (!e) return false;
-			if (!ui.selected.targets.length) return true;
-			if (!ui.selected.targets[0].countDiscardableCards(player, "ej", i => get.subtype(i) == "equip2")) {
-				return es.some(card => get.subtype(card) == "equip2") || js.some(card => get.subtype(card) == "equip2");
-			}
-			return true;
-		},
-		selectTarget: function() {
-			return 2;
-			if (!ui.selected.targets.length) return [1, 2];
-			let e = 0, player = get.event("player"), target = ui.selected.targets[0];
-			let es = target.getDiscardableCards(player, "e"), js = target.getDiscardableCards(player, "j", i => get.type(i) == "equip");
-			if (es.length) e++;
-			e += js.length;
-			if (e >= 2 && (es.some(card => get.subtype(card) == "equip2") || js.some(card => get.subtype(card) == "equip2"))) return [1, 2];
-			return 2;
-		},
-		complexTarget: true,
-		multitarget: true,
-		multiline: true,
-		async content(event, trigger, player) {
-			const targets = event.targets.slice();
-			if (targets.length == 1) {
-				await player.discardPlayerCard("ej", targets[0], true, 2).set("filterButton", button => {
-					let position = get.position(button.link), subtype = get.subtype(button.link);
-					if (!subtype || !subtype.startsWith("equip")) return false;
-					if (ui.selected.buttons.length) {
-						let pos = get.position(ui.selected.buttons[0].link), sub = get.subtype(ui.selected.buttons[0].link);
-						if (pos == "e" && position == "e") return false;
-						if (sub == "equip2") return true;
-						return subtype == "equip2";
-					}
-					if (position == "e") {
-						if (!get.event("js").some(i => get.subtype(i) == "equip2")) return subtype == "equip2";
-						return true;
-					}
-					if (!get.event("es").length) return subtype == "equip2";
-					return true;
-				}).set("es", targets[0].getDiscardableCards(player, "e", i => get.subtype(i) == "equip2")).set("js", targets[0].getDiscardableCards(player, "j", i => get.type(i) == "equip"));
-				return;
-			}
-			let canfj = targets.filter(target => {
-				return target.countDiscardableCards(player, "e", i => get.subtype(i) == "equip2");
-			});
-			for (let i = 0; i < 2; i++) {
-				if (i && canfj.includes(targets[i]) && !targets[i].countDiscardableCards(player, "e", i => get.subtype(i) == "equip2")) break;
-				const result = await player
-					.discardPlayerCard("e", targets[i], true)
-					.set("filterButton", button => {
-						if (get.event("fj")) return get.subtype(button.link) == "equip2";
-						return true;
-						return get.type(button.link) == "equip";
-					})
-					.set("fj", canfj.length === 1 && canfj.includes(targets[i]))
-					.forResult();
-				if (result.bool && get.subtype(result.cards[0]) == "equip2") {
-					canfj = [];
-				}
-			}
-		},
-		ai: {
-			order: 10,
-			result: {
-				target(player, target) {
-					let att = get.attitude(player, target), es = [];
-					target.countDiscardableCards(player, "e", i => {
-						es.push(get.value(i, target));
-					});
-					let min = Math.min(...es), max = Math.max(...es), ext = target.hasSkillTag("noe") ? 10 : 0;
-					if (att <= 0) return ext - max;
-					return ext - min;
-				}
-			},
-		},
-	},
-	stdqiansu: {
-		audio: "shengxi_feiyi",
-		trigger: { target: "useCardToTargeted" },
-		filter(event, player) {
-			return get.type2(event.card) == "trick" && !player.countCards("e");
-		},
-		frequent: true,
-		content() {
-			player.draw();
-		},
-		ai: {
-			noe: true,
-			effect: {
-				target(card, player, target) {
-					if (target.countCards("e")) return;
-					if (target == player && get.type(card) == "equip" && get.equipValue(card) < 5) return 0;
-					if (get.type2(card) == "trick") return [1, 0.6];
-				},
-			},
-		},
-	},
 	//标准版甘夫人
 	stdshushen: {
 		audio: "shushen",
