+++ character/standard/sort.js
@@ -3,7 +3,6 @@ const characterSort = {
 	standard_2013: ["old_re_lidian", "huaxiong", "re_yuanshu"],
 	standard_2019: ["gongsunzan", "xf_yiji"],
 	standard_2023: ["std_panfeng", "ganfuren"],
-	standard_shaoyin: ["std_sunhao", "std_mateng", "std_mayunlu", "std_jianggan", "std_zhouchu", "std_lvlingqi", "std_dc_yanghu", "std_dc_luotong", "std_lijue", "std_chengpu", "std_db_wenyang", "std_re_dengzhi", "std_zhangyì", "std_chengyu", "std_fanyufeng", "std_feiyi"],
 };
 
 const characterSortTranslate = {
@@ -11,7 +10,6 @@ const characterSortTranslate = {
 	standard_2013: "2013版标准包",
 	standard_2019: "2019版标准包",
 	standard_2023: "2023版标准包",
-	standard_shaoyin: "四象封印·少阴",
 };
 
 export { characterSort, characterSortTranslate };
