+++ character/standard/translate.js
@@ -179,84 +179,6 @@ const translates = {
 	stdshushen: "淑慎",
 	stdshushen_info: "当你回复1点体力时，你可以令一名其他角色摸一张牌（若其没有手牌则改为摸两张牌）。",
 	old_re_lidian: "李典",
-	std_sunhao: "标孙皓",
-	std_sunhao_prefix: "标",
-	std_mateng: "标马腾",
-	std_mateng_prefix: "标",
-	std_mayunlu: "标马云騄",
-	std_mayunlu_prefix: "标",
-	std_jianggan: "标蒋干",
-	std_jianggan_prefix: "标",
-	std_zhouchu: "标周处",
-	std_zhouchu_prefix: "标",
-	std_lvlingqi: "标吕玲绮",
-	std_lvlingqi_prefix: "标",
-	std_dc_yanghu: "标羊祜",
-	std_dc_yanghu_prefix: "标",
-	std_dc_luotong: "标骆统",
-	std_dc_luotong_prefix: "标",
-	std_lijue: "标李傕",
-	std_lijue_prefix: "标",
-	std_chengpu: "标程普",
-	std_chengpu_prefix: "标",
-	std_db_wenyang: "标文鸯",
-	std_db_wenyang_prefix: "标",
-	std_re_dengzhi: "标邓芝",
-	std_re_dengzhi_prefix: "标",
-	std_zhangyì: "标张翼",
-	std_zhangyì_prefix: "标",
-	std_chengyu: "标程昱",
-	std_chengyu_prefix: "标",
-	std_fanyufeng: "标樊玉凤",
-	std_fanyufeng_prefix: "标",
-	std_feiyi: "标费祎",
-	std_feiyi_prefix: "标",
-	stdcanshi: "残蚀",
-	stdcanshi_info: "锁定技，摸牌阶段，你改为摸X张牌（X为场上的已受伤角色且X至少为1）。然后本回合你使用【杀】或普通锦囊牌指定目标后，若其已受伤，你弃置一张牌。",
-	stdxiongyi: "雄异",
-	stdxiongyi_info: "限定技，出牌阶段，你可以选择任意名角色，这些角色依次选择是否使用一张不可被响应的【杀】，然后这些角色重复此流程直至有角色不使用【杀】。",
-	stdyouji: "游骑",
-	stdyouji_info: "主公技，准备阶段，你可以移动一名群势力角色的一张坐骑牌。",
-	stdfengpo: "凤魄",
-	stdfengpo_info: "当你使用【杀】造成伤害时，你可以弃置你或其的一张牌，若以此法弃置了方片牌，则此伤害+1。",
-	stddaoshu: "盗书",
-	stddaoshu_info: "每轮限一次，一名角色的准备阶段，你可以展示除其外一名角色的一张牌，然后令其获得此牌，且你与其本回合不能使用与此牌花色相同的牌。",
-	stddaizui: "戴罪",
-	stddaizui_info: "锁定技，当你受到伤害后，你视为本轮未发动过〖盗书〗。",
-	stdxiongxia: "凶侠",
-	stdxiongxia_info: "你可以将两张牌当作【决斗】对两名其他角色使用。你以此法使用的【决斗】结算完毕后，若所有目标角色都受到了此牌造成的伤害，则〖凶侠〗于本回合失效。",
-	stdhuizhan: "挥戟",
-	stdhuizhan_info: "你使用【杀】可以额外指定至多两个目标。若如此做，目标角色响应此【杀】时，其他目标角色可以代替其使用【闪】。",
-	stdmingfa: "明伐",
-	stdmingfa_info: "出牌阶段，你可以对一名体力值大于1的角色造成1点伤害，然后此技能失效直至其死亡或回复体力。",
-	stdjinjian: "进谏",
-	stdjinjian_info: "每回合每项各限一次，当你造成/受到伤害时，你可令此伤害+1/-1，然后你本回合内下一次造成的伤害-1/+1且不能触发〖进谏〗。",
-	stdrenzheng: "仁政",
-	stdrenzheng_info: "锁定技，当有伤害被防止时，你令当前回合角色摸一张牌。",
-	stdxiongsuan: "凶算",
-	stdxiongsuan_info: "锁定技，准备阶段，若你的体力值为全场最多，则你须对至少一名体力值等于你的角色各造成1点伤害。",
-	stdchunlao: "醇醪",
-	stdchunlao_info: "弃牌阶段结束时，若你本阶段弃置了不少于两张牌，则你可以用这些牌交换一名其他角色的手牌，然后其可以令你回复1点体力。",
-	stdquedi: "却敌",
-	stdquedi_info: "你可以将【杀】当作【决斗】使用。",
-	stdzhiyinmeng: "急盟",
-	stdzhiyinmeng_info: "准备阶段，你可以交给一名其他角色任意张牌，然后其可以交给你任意张牌。",
-	stdhehe: "和合",
-	stdhehe_info: "摸牌阶段结束时，你可以令至多两名手牌数与你相同的其他角色各摸一张牌。",
-	stdzhiyi: "执义",
-	stdzhiyi_info: "锁定技，一名角色的回合结束时，若你本回合使用过【杀】，则你视为使用【杀】或摸一张牌。",
-	stdshefu: "设伏",
-	stdshefu_info: "①结束阶段，你可以将一张手牌称为“伏兵”扣置于武将牌上。②一名角色使用牌时，你可以移去武将牌上的一张与此牌同名的“伏兵”并令此牌无效。",
-	stdyibing: "益兵",
-	stdyibing_info: "一名角色进入濒死状态时，你可以获得其一张牌。",
-	stdbazhan: "把盏",
-	stdbazhan_info: "出牌阶段限一次，你可以交给一名男性角色一张手牌，然后其可以交给你一张与此牌类别不同的牌。",
-	stdzhanying: "醮影",
-	stdzhanying_info: "锁定技，你的回合内，手牌数比回合开始时多的角色不能使用红色牌且受到的伤害+1。",
-	stdtiaohe: "调和",
-	stdtiaohe_info: "出牌阶段限一次，你可以弃置场上的一张装备牌和一张防具牌（不能为同一名角色的牌）。",
-	stdqiansu: "谦素",
-	stdqiansu_info: "当你成为锦囊牌的目标后，若你的装备区没有牌，则你可以摸一张牌。",
 };
 
 export default translates;
