+++ character/yijiang/characterReplace.js
@@ -3,14 +3,14 @@ const characterReplaces = {
 	zhangchunhua: ["zhangchunhua", "re_zhangchunhua", "ol_zhangchunhua", "jin_zhangchunhua", "star_zhangchunhua"],
 	yujin: ["yujin", "yujin_yujin", "ol_yujin", "sb_yujin", "xin_yujin", "re_yujin"],
 	dc_xushu: ["re_xushu", "dc_xushu"],
-	xushu: ["xin_xushu", "xushu"],
+	xushu: ["xin_xushu", "xushu", "std_xushu"],
 	fazheng: ["xin_fazheng", "ol_fazheng", "re_fazheng", "sb_fazheng", "tw_re_fazheng", "fazheng"],
 	masu: ["xin_masu", "re_masu", "masu"],
 	xusheng: ["xusheng", "xin_xusheng", "re_xusheng", "old_xusheng"],
 	wuguotai: ["wuguotai", "xin_wuguotai", "re_wuguotai", "ol_wuguotai"],
 	lingtong: ["lingtong", "xin_lingtong", "ol_lingtong", "re_lingtong", "old_lingtong"],
 	gaoshun: ["gaoshun", "xin_gaoshun", "ol_gaoshun", "re_gaoshun", "sb_gaoshun", "old_gaoshun"],
-	zhonghui: ["zhonghui", "xin_zhonghui", "re_zhonghui", "old_zhonghui", "pe_zhonghui", "clan_zhonghui", "yj_zhonghui"],
+	zhonghui: ["zhonghui", "xin_zhonghui", "re_zhonghui", "old_zhonghui", "pe_zhonghui", "clan_zhonghui", "yj_zhonghui", "std_zhonghui"],
 	wangyi: ["wangyi", "ol_wangyi", "re_wangyi", "old_wangyi"],
 	caozhang: ["caozhang", "ol_caozhang", "re_caozhang", "xin_caozhang"],
 	guanzhang: ["guanzhang", "re_guanzhang", "old_guanzhang"],
@@ -19,7 +19,7 @@ const characterReplaces = {
 	bulianshi: ["bulianshi", "dc_bulianshi", "re_bulianshi", "old_bulianshi"],
 	handang: ["handang", "xin_handang", "re_handang", "tw_handang", "sb_handang", "old_handang"],
 	chengpu: ["chengpu", "ol_chengpu", "re_chengpu", "xin_chengpu", "tw_chengpu", "ns_chengpu", "std_chengpu"],
-	liubiao: ["liubiao", "xin_liubiao", "re_liubiao", "jsrg_liubiao", "sb_liubiao", "oldre_liubiao", "old_liubiao", "ol_liubiao"],
+	liubiao: ["liubiao", "xin_liubiao", "re_liubiao", "jsrg_liubiao", "sb_liubiao", "oldre_liubiao", "old_liubiao", "ol_liubiao", "std_liubiao"],
 	manchong: ["manchong", "re_manchong"],
 	caochong: ["caochong", "re_caochong", "old_caochong", "ol_caochong"],
 	guohuai: ["guohuai", "re_guohuai", "xin_guohuai", "tw_guohuai", "ol_guohuai"],
@@ -28,7 +28,7 @@ const characterReplaces = {
 	yufan: ["yufan", "ol_yufan", "xin_yufan", "re_yufan"],
 	zhuran: ["zhuran", "re_zhuran", "xin_zhuran", "old_zhuran"],
 	liru: ["xin_liru", "liru", "dc_liru", "re_liru", "yj_liru", "ol_liru"],
-	fuhuanghou: ["fuhuanghou", "re_fuhuanghou", "xin_fuhuanghou", "old_fuhuanghou"],
+	fuhuanghou: ["fuhuanghou", "re_fuhuanghou", "xin_fuhuanghou", "old_fuhuanghou", "std_fuhuanghou"],
 	chenqun: ["chenqun", "dc_chenqun", "re_chenqun", "old_chenqun"],
 	hanhaoshihuan: ["hanhaoshihuan", "re_hanhaoshihuan"],
 	caozhen: ["caozhen", "re_caozhen", "xin_caozhen", "old_caozhen"],
@@ -45,7 +45,7 @@ const characterReplaces = {
 	zhangrang: ["zhangrang", "ol_zhangrang", "junk_zhangrang"],
 	jikang: ["jikang", "re_jikang", "dc_jikang"],
 	xinxianying: ["xinxianying", "re_xinxianying", "ol_xinxianying", "sp_xinxianying"],
-	gongsunyuan: ["gongsunyuan", "re_gongsunyuan"],
+	gongsunyuan: ["gongsunyuan", "re_gongsunyuan", "std_gongsunyuan"],
 	zhoucang: ["zhoucang", "re_zhoucang", "xin_zhoucang"],
 	guotufengji: ["guotufengji", "re_guotufengji"],
 	guanping: ["guanping", "re_guanping", "dc_sb_guanping"],
@@ -59,7 +59,7 @@ const characterReplaces = {
 	caiyong: ["caiyong", "re_caiyong", "yue_caiyong"],
 	chengong: ["chengong", "re_chengong", "sb_chengong"],
 	xunyou: ["xunyou", "re_xunyou", "clan_xunyou"],
-	xuezong: ["xuezong", "tw_xuezong"],
+	xuezong: ["xuezong", "tw_xuezong", "std_xuezong"],
 	huanghao: ["huanghao", "dc_huanghao", "old_huanghao"],
 	caorui: ["caorui", "re_caorui", "old_caorui"],
 	sunziliufang: ["sunziliufang", "dc_sunziliufang"],
